requirements.txt
pandas==2.2.0
numpy==1.26.4

SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

python-dotenv==1.0.1

tqdm==4.66.2
loguru==0.7.2
prefect